% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_years.R
\name{activity_years}
\alias{activity_years}
\title{Visualise google search activity over the course of several years for different subjects}
\usage{
activity_years(dataset, activity)
}
\arguments{
\item{dataset}{Dataset including multiple years, month and certain activities}

\item{activity}{What activity you want to visualise}
}
\value{
A plot
}
\description{
Visualise google search activity over the course of several years for different subjects
}
\examples{

dataset <- data.frame(year=(2002:2013),month=(1:12),activity1=(12:1),activity2=(24:35))
activity_years(dataset,"activity1")
activity_years(dataset,"activity2")
}
