% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_years_func.R
\name{activity_years_func}
\alias{activity_years_func}
\title{Visualize google search activity over the course of several years for different diseases}
\usage{
activity_years_func(dataset, name)
}
\arguments{
\item{dataset}{Dataset including multiple years, month and certain activities}

\item{name}{What disease you want to visualize}
}
\value{
A plot
}
\description{
Visualize google search activity over the course of several years for different diseases
}
\examples{

dataset <- data.frame(year=(2002:2013),
                      month=(1:12),
                      activity=(12:1),
                      disease=(rep(c("flu","dengue"),times=6)))
activity_years_func(dataset,"flu")
activity_years_func(dataset,"dengue")
}
