% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_progressplot.R
\name{covid19_progressplot}
\alias{covid19_progressplot}
\title{Function for visulaising covid cases and deaths}
\usage{
covid19_progressplot(dataset, condition)
}
\arguments{
\item{dataset}{Filtered dataset with covid data}

\item{condition}{Choose cases or deaths}
}
\value{
A plot with date range on x-axis and amount of deaths/cases on y axis
}
\description{
Function for visulaising covid cases and deaths
}
\examples{
dataset <- data.frame(date_range=(20200512:20200812),
           deaths=(1:301),cases=(1:301),
           dateRep=(20200512:20200812))
covid19_progressplot(dataset,"cases")
covid19_progressplot(dataset,"deaths")
}
