% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_gapminder_netherlands.R
\docType{data}
\name{gapminder_netherlands}
\alias{gapminder_netherlands}
\title{Health and income of the Netherlands}
\format{
A data frame with 10 rows and 7 variables:
\describe{
\item{country}{country}
\item{year}{year}
\item{population}{Population of country}
\item{gdp}{According to World Bankdev}
\item{infant_mortality }{Infant deaths per 1000}
\item{life_expectancy }{In years}
\item{fertility}{Average number of children per woman}
}
}
\source{
\url{https://cran.r-project.org/web/packages/dslabs/index.html}
}
\usage{
gapminder_netherlands
}
\description{
A dataset containing information about the health and income of the Netherlands between 2002 and 20011
}
\keyword{datasets}
