% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_func.R
\name{population_func}
\alias{population_func}
\title{Plots with means of variables of  population, GDP from gapminder and activity}
\usage{
population_func(dataset, year, name)
}
\arguments{
\item{dataset}{all_data dataset}

\item{year}{Year you want data about}

\item{name}{Column/disease you want to know the mean activity of}
}
\value{
A stacked plot comparing population, gdp and activity
}
\description{
Plots with means of variables of  population, GDP from gapminder and activity
}
\examples{
dataset <- data.frame(year=(2002:2013),
country=letters[1:12],
population=(1:12),
gdp=(1:12),
month=(1:12),
activity=(12:1),
disease=(rep(c("flu","dengue"),times=6)))

population_func(dataset,2008,"flu")
population_func(dataset,2008,"dengue")
}
