% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_func.R
\name{tidy_func}
\alias{tidy_func}
\title{Function used for making datasets tidy and splitting date and adding disease to table}
\usage{
tidy_func(path, name)
}
\arguments{
\item{path}{path to file}

\item{name}{Disease choosen to make tidy}
}
\value{
A tidy dataset
}
\description{
Function used for making datasets tidy and splitting date and adding disease to table
}
\examples{
test <- c(rep(c(" , , , "), 11),"Date,a,b,c","2002-12-29,1,2,3", "2003-01-05,4,5,6",
"2003-01-12,7,8,9")

tf <-tempfile()
writeLines(test,tf)
tidy_func(tf,"dengue")
unlink(tf)

}
