% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_func.R
\name{population_func}
\alias{population_func}
\title{Summarises means of variables of  population, GDP from gapminder and activity life expectancy from the google dataset an dmake a plot}
\usage{
population_func(dataset, year, name)
}
\arguments{
\item{dataset}{all_data dataset}

\item{year}{year you want data about}

\item{name}{column/disease you want the mean of}
}
\value{
A summarised dataset
}
\description{
Summarises means of variables of  population, GDP from gapminder and activity life expectancy from the google dataset an dmake a plot
}
\examples{
dataset <- data.frame(year=(2002:2013),
country=letters[1:12],
population=(1:12),
gdp=(1:12),
month=(1:12),
activity=(12:1),
disease=(rep(c("flu","dengue"),times=6)))

population_func(dataset,2008,"flu")
population_func(dataset,2008,"dengue")
}
