% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_func.R
\name{tidy_func}
\alias{tidy_func}
\title{Function used for making datasets tidy and splitting date and adding disease to table}
\usage{
tidy_func(path, name)
}
\arguments{
\item{path}{path to file}

\item{name}{disease choosen to make tidy}
}
\value{
A tidy dataset
}
\description{
Function used for making datasets tidy and splitting date and adding disease to table
}
